fix: simplify error handling by normalizing pocolog files in groups
This commit uses the (valid) assumption that a data stream,
once encountered, will always be logged by the same logger in the
same "file group" (that is, the same sequence of rotated logs).

So, if we process all log files of a group `NAME.SEQ.log` together,
recovering from an error is simply removing all output files created
during the processing. This removes the need for a log of bookkeeping.