chore: simplify CLI::Datastore
Moved reporter as an instance variable, as it is part of the general
context, and was essentially passed to every single method. Moreover,
now that Import only takes care of the import into a core/cache path
pair (not considering the datastore itself), remove now-obsolete methods
in Import and cleanup CLI::Datastore accordingly.