chore: remove the export to pocolog indexes during import
With the introduction of the minimal index files, we need those
only when reading a file, and one usually does not read the majority
of log files.

Additionally, it was not actually saved with compressed files (since
we have anyways to decompress the files to use them), and was potentially
using a lot of memory during import.