feat: allow to optionally ignore incompatible types
The nature of Typelib logging is that when changing architectures,
or because of type changes, old log files cannot necessarily replayed
as-is.

While pocolog has support to "upgrade" the data in simple (common)
cases, it's currently broken on `syskit replay`. For now, provide
a way to simply ignore these streams.