fix: explicitly use syskit models in replay deployment definitions
Syskit was until now expecting an orogen model, which was then
re-resolved to a Syskit model "under the hood". This was messed
up by the fact that in syskit-log we define more than one syskit
model for a given orogen model.

Explicitly pass syskit models for clarity and comment where things
are still a bit ... shaky.

This fixes running service-narrowed task streams (the value returned by
TaskStreams#as_data_service)